Simplify service manager options and merge endpoints from configuration. (#1111)
* read endoints array in ServiceManagerOptions from IConfiguration

* move validate options from ServiceManagerOptions to ServiceManagerContext
