+++ src/Microsoft.Azure.SignalR.Common/Endpoints/IServiceEndpointOptionExtensions.cs
@@ -1,0 +1,36 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System.Collections.Generic;
+
+namespace Microsoft.Azure.SignalR.Common.Endpoints
+{
+    internal static class IServiceEndpointOptionExtensions
+    {
+        public static IEnumerable<ServiceEndpoint> GetMergedEndpoints(this IServiceEndpointOptions options)
+        {
+            if (options == null)
+            {
+                yield break;
+            }
+
+            var endpoints = options.Endpoints;
+            var connectionString = options.ConnectionString;
+
+            if (!string.IsNullOrEmpty(connectionString))
+            {
+                yield return new ServiceEndpoint(options.ConnectionString);
+            }
+
+            // ConnectionString can be set by custom Configure
+            // Return both the one from ConnectionString and from Endpoints
+            if (endpoints != null)
+            {
+                foreach (var endpoint in endpoints)
+                {
+                    yield return endpoint;
+                }
+            }
+        }
+    }
+}
