+++ src/Microsoft.Azure.SignalR.Management/Configuration/ServiceManagerContext.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System;
 using System.Net;
 
 namespace Microsoft.Azure.SignalR.Management
@@ -20,5 +21,13 @@ namespace Microsoft.Azure.SignalR.Management
         public ServiceTransportType ServiceTransportType { get; set; } = ServiceTransportType.Transient;
 
         public bool DisposeServiceProvider { get; set; } = false;
+
+        internal void ValidateOptions()
+        {
+            if (ServiceEndpoints.Length == 0)
+            {
+                throw new InvalidOperationException($"Service endpoint(s) is/are not configured.");
+            }
+        }
     }
 }
