+++ src/Microsoft.Azure.SignalR.Management/Configuration/ServiceManagerContextSetup.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System.Linq;
+using Microsoft.Azure.SignalR.Common.Endpoints;
 using Microsoft.Azure.SignalR.Management.Configuration;
 using Microsoft.Extensions.Options;
 
@@ -14,9 +16,7 @@ namespace Microsoft.Azure.SignalR.Management
 
         protected override void Convert(ServiceManagerContext target, ServiceManagerOptions source)
         {
-            target.ServiceEndpoints = source.ServiceEndpoints ?? (source.ServiceEndpoint != null
-                    ? (new ServiceEndpoint[] { source.ServiceEndpoint })
-                    : (new ServiceEndpoint[] { new ServiceEndpoint(source.ConnectionString) }));
+            target.ServiceEndpoints = source.GetMergedEndpoints().ToArray();
             target.ApplicationName = source.ApplicationName;
             target.ConnectionCount = source.ConnectionCount;
             target.Proxy = source.Proxy;
