+++ src/Microsoft.Azure.SignalR.Management/Configuration/ServiceManagerOptions.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
 using System.Net;
 
 namespace Microsoft.Azure.SignalR.Management
@@ -9,8 +8,10 @@ namespace Microsoft.Azure.SignalR.Management
     /// <summary>
     /// Configurable options for Azure SignalR Management SDK.
     /// </summary>
-    public class ServiceManagerOptions
+    public class ServiceManagerOptions : IServiceEndpointOptions
     {
+        private ServiceEndpoint[] _serviceEndpoints;
+
         /// <summary>
         /// Gets or sets the ApplicationName which will be prefixed to each hub name
         /// </summary>
@@ -32,66 +33,15 @@ namespace Microsoft.Azure.SignalR.Management
         public IWebProxy Proxy { get; set; }
 
         /// <summary>
-        /// Gets or sets a service endpoint of Azure SignalR Service.
-        /// </summary>
-        internal ServiceEndpoint ServiceEndpoint { get; set; }
-
-        /// <summary>
         /// Sets multiple service endpoints of Azure SignalR Service.
         /// </summary>
-        internal ServiceEndpoint[] ServiceEndpoints { get; set; } //not ready for public use
+        ServiceEndpoint[] IServiceEndpointOptions.Endpoints => _serviceEndpoints;  //todo not ready for public use
+
+        internal ServiceEndpoint[] Endpoints { get => _serviceEndpoints; set => _serviceEndpoints = value; }
 
         /// <summary>
         /// Gets or sets the transport type to Azure SignalR Service. Default value is Transient.
         /// </summary>
         public ServiceTransportType ServiceTransportType { get; set; } = ServiceTransportType.Transient;
-
-        /// <summary>
-        /// Method called by management SDK to validate options.
-        /// </summary>
-        internal void ValidateOptions()
-        {
-            ValidateServiceEndpoint();
-            ValidateServiceTransportType();
-        }
-
-        private void ValidateServiceEndpoint()
-        {
-            var notNullCount = 0;
-            if (ConnectionString != null)
-            {
-                notNullCount += 1;
-            }
-            if (ServiceEndpoint != null)
-            {
-                notNullCount += 1;
-            }
-            if (ServiceEndpoints != null)
-            {
-                notNullCount += 1;
-            }
-
-            if (notNullCount == 0)
-            {
-                throw new InvalidOperationException($"Service endpoint(s) is/are not configured. Please set one of the following properties {nameof(ConnectionString)}, {nameof(ServiceEndpoint)}, {nameof(ServiceEndpoints)}.");
-            }
-            if (notNullCount > 1)
-            {
-                throw new InvalidOperationException($"Please set ONLY one of the following properties: {nameof(ConnectionString)}, {nameof(ServiceEndpoint)}, {nameof(ServiceEndpoints)}.");
-            }
-            if (ServiceEndpoints != null && ServiceEndpoints.Length == 0)
-            {
-                throw new InvalidOperationException($"The length of parameter {nameof(ServiceEndpoints)} is zero.");
-            }
-        }
-
-        private void ValidateServiceTransportType()
-        {
-            if (!Enum.IsDefined(typeof(ServiceTransportType), ServiceTransportType))
-            {
-                throw new ArgumentOutOfRangeException($"Not supported service transport type. " +
-                    $"Supported transport types are {ServiceTransportType.Transient} and {ServiceTransportType.Persistent}.");
-            }
-        }
     }
 }
