+++ src/Microsoft.Azure.SignalR.Management/Configuration/ServiceManagerOptionsSetup.cs
@@ -24,6 +24,9 @@ namespace Microsoft.Azure.SignalR.Management
             if (_configuration != null)
             {
                 _configuration.GetSection(Constants.Keys.AzureSignalRSectionKey).Bind(options);
+
+                //Our configuration format of service endoints array is not the standard format to configure array supported by .NET
+                options.Endpoints = _configuration.GetSignalRServiceEndpoints(Constants.Keys.ConnectionStringDefaultKey);
             }
         }
